refactor: 🛠 move about page piece to Component directory
