+++ src/components/AboutPage/index.jsx
@@ -1,0 +1,5 @@
+import ImageGallery from './ImageGallery';
+import InfoSection from './InfoSection';
+import styled from './styled';
+
+export { ImageGallery, InfoSection, styled };
